Refactor `TwakeDialog`
